фÐfeat(home-page-design) set-up a template for gallery and add motivational text
