+++ be-a-human/src/App.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { RouterLink, RouterView } from 'vue-router'
+import 'primeicons/primeicons.css'
 </script>
 
 <template>
