+++ package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "@primevue/themes": "^4.2.3",
+    "primeicons": "^7.0.0",
     "primevue": "^4.2.3"
   }
 }
